Create purge-files-smb-advanced
This script is used to delete old files from a specified SMB mount folder and log the deleted and error files. 

The script starts by defining the smb_path variable which is the path to the SMB mount folder. It then gets the current time in seconds and the disk usage for the SMB mount folder. It creates two lists (deleted_files and error_files) to store the deleted and error files. 

The script then defines a process_file function to process each file. This function checks if the file is older than one day and if so, it deletes the file. If there is an error when deleting the file, it is added to the error_files list and an error message is printed.

The script then checks the disk usage. If the disk usage is above 90%, it recursively traverses the folder and processes each file. If the disk usage is below 90%, it still recursively traverses the folder and processes each file. 

If the disk usage is above 60%, it recursively traverses the folder and processes each file in order of oldest to newest. It then checks the disk usage after each file is processed and exits the loop

This script requires Python 3.6 or higher and the following libraries to be installed: os, time, shutil, concurrent.futures, and daemon.

The use of Multithreading: ThreadPoolExecutor is used to process the files and folders in parallel. This should significantly speed up the script, especially if the folder contains a large number of files and folders.

You can also remove the daemonizer and run it as a systemctl service : 

In order to run your script as a systemctl service, you will need to create a service file that tells systemctl how to run your script. The service file is a text file with a specific format and a specific location on your system.

Here are the steps to create a service file and run your script as a systemctl service:

1. Create a new file in the /etc/systemd/system directory. The file name should end in .service, for example my_script.service.

2. Open the file in a text editor and add the following content:

[Unit]
Description=My Script

[Service]
ExecStart=/usr/bin/python3 /path/to/script.py
Restart=always
User=username

[Install]
WantedBy=multi-user.target


- 'ExecStart' is the path to the Python interpreter and the path to your script.
- 'User' is the username of the user that the script should run as.
- 'Restart' is used to configure if the service should be restarted if it crashes or fails.

3. Save the file and exit the text editor.

4. Run the following command to reload the systemctl configuration and enable the service:

sudo systemctl daemon-reload
sudo systemctl enable my_script.service

5. Finally, you can start/stop or check status of the service by running the following command:
sudo systemctl start my_script.service
sudo systemctl stop my_script.service
sudo systemctl status my_script.service