Create pan-create-objects
Palo Alto Panorama Object Adder
This script allows you to add new IP-Netmask objects to a Palo Alto Panorama device, given a CSV file containing the current and new objects to be added. The script connects to the Palo Alto Panorama device using the pan.xapi module, reads the objects from the CSV file, checks if the current and new objects already exist, and if not, creates the new objects in the same device-group as the current objects. The script uses the progressbar module to display a progress bar for the operation.

How to use:
Install the required modules by running pip install pan-python progressbar in your terminal.
Run the script by executing python script.py.
Enter your credentials (username, password), the address of your Palo Alto Panorama device, and the name of your CSV file when prompted.
The script will read the objects from the CSV file, check if the current and new objects already exist, and create new objects as necessary.
A progress bar will be displayed, showing the progress of the operation.

CSV file format:
The CSV file should contain two columns, with the current object in the first column and the new object in the second column. The first row should be a header row. An example file is shown below:

------------------------------
current_object,new_object
10.0.0.0/24,10.0.1.0/24
192.168.0.0/24,192.168.1.0/24
------------------------------

Exceptions:
The script will catch any exceptions that occur and print an error message. If an error occurs, the script will stop executing and the changes will not be committed to the Palo Alto Panorama device.

Requirements:
This script requires the following modules to be installed:

pan-python
progressbar

You can install these modules by running pip install pan-python progressbar in your terminal.

__________________________________________________

This script is written in Python and uses the pan-python library to connect to a Palo Alto Panorama firewall. The script takes in the following inputs:

username: the username to log into the Palo Alto Panorama
password: the password to log into the Palo Alto Panorama
address: the address of the Palo Alto Panorama
filename: the name of a csv file (including the .csv extension) that contains a list of objects to be created or checked
The script performs the following tasks:

Connects to the Palo Alto Panorama using the inputs provided
Opens the csv file and reads in each row
For each row, the script checks if the current object exists in the Palo Alto Panorama by sending an API call
If the current object does not exist, it skips the line and moves on to the next row
If the current object exists, it gets the device-group of the current object and checks if the new object already exists in the Palo Alto Panorama by sending another API call
If the new object does not exist, it creates the new object in the same device-group as the current object
If the new object already exists, it skips it and moves on to the next row
After all rows have been processed, the script commits all changes made to the Palo Alto Panorama
The output of the script shows the result of each operation, such as if a new object was created or if an object already exists. The script will print a message indicating whether the current object exists, whether the new object was created, or whether the new object already exists.